05-open-data/sensors/requirements.txt
python-dotenv
requests

pip~=21.1
setuptools~=56.0.0
